Add files via upload
This is the code that I used to create the plots for the final presentation (currently on slides 21-23). I know it's in python (The Forbidden Fruit), but it was just what was quickest for me. Also, I think we're supposed to use python in the Spring 2022 semester, but I don't know.